+++ playwright/tests/kompassiLogin.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { populateDb, postSettings } from "playwright/utils";
+import { populateDb, postSettings } from "playwright/playwrightUtils";
 import { LoginProvider } from "shared/config/eventConfigTypes";
 
 test("Kompassi login", async ({ page, request }) => {
